+++ CHANGES.rst
@@ -17,6 +17,9 @@ Unreleased
   - Moved IRC server and bot name configuration from command line
     arguments to configuration file.
 
+  - Moved HTTP receiver configuration from command line arguments to
+    configuration file.
+
   - Made configuration filename a required command line argument.